+++ CHANGELOG.md
@@ -15,6 +15,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Add shorthand `cx.Dropzone('dropzoneId', [require('elem1'), require('elem2')])`
 * Update Twing from 5.x to 7.x
 * Add handlebars functions to twig
+* Add `withExtendAllDropzones()` to content-part and template-part
 
 
 ## [2.0.1] - 03.12.2025